Implement info_capture as a scoped holder of informative messages
This is supposed to be used both for runtime and compile time
evaluation.

The split_expression_string() is required because the desired interface,
CAPTURE(var, var2, ...) doesn't allow for stringification of individual
arguments. Instead, __VA_ARGS__ can be stringized as a whole, and then
split with this function.

info_capture_stringification_helper() is a small helper function to
generate info strings of the form "variable_name: variable_content".