Fix xml_writer throwing instead of asserting
Either way, it would abort - previously, because the xml_reporter is
noexcept and didn't catch the exception. Now because of the assertion.
But with the assertion, at least it aborts at the place of error.

In any case, if this assertion fires, that indicates a logic error and
needs to be fixed. The user cannot do anything about it.