Improve constexpr error reporting
Now the errors are not limited to the failed assertion, the expansion of
the failed expression, and an additional message, but rather to any
number of messages that fit within the maximum length (currently 1024
characters).

This changeset also singles out the expansion as a separate parameter
to reporters instead of one of any number of messages. This makes it
easier for reporters to separately report it, and it also allows for
cases without any decomposition.