Implement runtime_if() and compiletime_if() constraints for SECTIONs
These constraints can be used to make the evaluation depend on a
compile-time constant boolean expression. This can be useful to stay
forward-compatible: If a unit can not be tested at compile time yet, but
there is a concept, trait or macro available that evaluates to whether
or not it can, then this can be used to make the section runtime-only
in the time being.